Do not set tqma8xqp machine as imx
Although based on an iMX SoC, the iMX8X series SoC is sufficiently
different from existing iMX ports that it does not make sense for
it to be considered as an 'imx' platform. Specifically, the fact
that the iMX8X uses a SCU micrcontroller with SCFW firmware to
control clocks, and the fact that the UARTs are low-power UARTs
with different register layout to the other iMX SoC UARTs make it
unhelpful to consider this specific platform an 'imx'.