Fix MCS refill_order assertion on SMP
The `refill_order` assertion is only applicable when the scheduling
context is not round-robin.

Assertions are changed such that `refill_order` is only asserted
when the the scheduling context is *not* a round-robin scheduling
context.

Additionally, in `maybe_add_empty_tail` the `empty_tail` element,
which is used to track time, is now explicitly all zeroes, rather
than unnecessarily initiliazing the `rTime` field. The only time
that `rTime` field was used for round-robin scheduling contexts
previously was in the `refill_order` check (which is not not
called for round-robin scheduling contexts).

See: https://github.com/seL4/seL4/issues/262

Signed-off-by: Ben Leslie <benno@brkawy.com>