Update gicv3 driver with basic hyp support
This is unlikely to be a working version of the virtual
interrupt controller, however it is sufficient to get through
bootstrap.

This is a useful improvement as it means that testing a kernel
running in EL2 mode is possible. This is a first step towards
getting a fully working GICv3.