Disable Vtimer update voffset option
This option (which is otherwise enabled by default) breaks Linux
time-keeping.

This could be seen easily by performing a "sleep 10" in Linux while
counting wall-clock time with a stop-watch.

With this option enabled (the default) the seL4 kernel adjusts the
virtual timer (which is used by Linux for timekeeping) so that it
is effectively only active while the Linux kernel is actively running.
This leads to the Linux kernel effectively "losing" time.

With this option enabled wall-clock time is accurately reported to
Linux.