Refactor kernel sources into 'kernel' directory
The seL4 repo contains both the kernel and libsel4 sources.
Both kernel and libsel4 have src and include directories as well
as build scripts. Having the kernel's src/include/build scripts in
the root, while libsel4 is in a sub-directory is not ideal.

This makes it difficult to have a single build script for building
both (which is a specific goal of sel4core effort).