Add initial script for generating SDK output
This is an initial proof-of-concept. Ideally, the overall
set up would be one in which there was a single build system
that generated the output SDK, rather than a script on top of
the build-system! But for practical purposes this ends up being
the best approach (at least in terms of getting a proof-of-concept
up and running)