Along normal memory to be reserved by the startup protocol (ARM only)
It is very useful to have the loader initialize some memory that is
neither the kernel, nor the initial task's image.

This is memory that the initial task should be able to turn into
frames and ultimately map into an appropriate virtual address
space.

Non-device-memory is always zeroed by the kernel before allowing
access from the user code.

To support the use case the semantics for the last two registers
in the startup protocol are altered. Rather than providing a
device-tree-blob region, these registers now define an
"extra device memory region".

This region shall be covered by device untyped objects. This
allows for the loader to pass a region of memory that can be
directly used by the init thread. Ideally, multiple regions
would be supported, however that would require more extensive
reworking of the startup protocol.